requirements.txt
django==4.2.11; python_version >= '3.8'
django-appconf==1.0.6; python_version >= '3.7'
django-axes==6.4.0; python_version >= '3.7'
django-ckeditor==6.7.1; python_version >= '3.8'
django-cors-headers==4.3.1; python_version >= '3.8'
django-debug-toolbar==4.3.0; python_version >= '3.8'
django-environ==0.10.0; python_version >= '3.5' and python_version < '4'
django-extensions==3.2.3; python_version >= '3.6'
django-imagekit==5.0.0
django-import-export==3.3.9; python_version >= '3.8'
django-js-asset==2.2.0; python_version >= '3.8'
django-jsonform==2.22.0; python_version >= '3.4'
django-polymorphic==3.1.0
django-redis==5.4.0; python_version >= '3.6'
django-tree-queries==0.19.0; python_version >= '3.8'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
redis==5.0.4; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
unidecode==1.3.8; python_version >= '3.5'
urllib3==2.2.1; python_version >= '3.8'
gunicorn==21.2.0; python_version >= '3.5'
setuptools==69.5.1; python_version >= '3.8'
virtualenv==20.26.1; python_version >= '3.7'
jdatetime
django-filter